Fix int/long build errors on Mac.
On Mac, casting a long to an int is a truncation from a 64-bit signed 
integer to a 32-bit signed integer.  This patch fixes some compile 
warnings/errors.  This is not a problem on Windows because an int and
a long are both 32-bit signed integers.  This diff fixes the build.
See http://trac.webkit.org/browser/trunk/Source/ThirdParty/ANGLE/changes.diff

Patch authored by achristensen07

BUG=angle:635

+++ src/common/mathutil.h
@@ -478,7 +478,7 @@ inline unsigned int average(unsigned int a, unsigned int b)
 
 inline signed int average(signed int a, signed int b)
 {
-    return ((long long)a + (long long)b) / 2;
+    return (signed int)((long long)a + (long long)b) / 2;
 }
 
 inline float average(float a, float b)
